Finished Lower Service class

+++ src/main/java/com/awoodcock/armory/service/ar/LowerService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,18 +29,64 @@ public class LowerService {
     }
 
     // READ
-    public Optional<Lower> getLowerById (int id) {
-        return lowerRepository.findById(id);
+    public Lower getLowerById(int id) {
+        Optional<Lower> lower = lowerRepository.findById(id);
+        return lower.orElse(null);
     }
 
-    public List<Lower> getAllLowers () {
+    public List<Lower> getAllLowers() {
         return lowerRepository.findAll();
     }
 
     // UPDATE
+    public Lower updateLowerById(int id, Lower lower) {
+        Optional<Lower> lowerOptional = lowerRepository.findById(id);
+
+        if (lowerOptional.isPresent()) {
+            Lower originalLower = lowerOptional.get();
+
+            // name
+            if (Objects.nonNull(lower.getName()) && !"".equalsIgnoreCase(lower.getName())) {
+                originalLower.setName(lower.getName());
+            }
+
+            // description
+            if (Objects.nonNull(lower.getDescription()) && !"".equalsIgnoreCase(lower.getName())) {
+                originalLower.setDescription(lower.getDescription());
+            }
+
+            // price
+            if (Objects.nonNull(lower.getPrice()) && lower.getPrice() != 0) {
+                originalLower.setPrice(lower.getPrice());
+            }
+
+            // weight
+            if (Objects.nonNull(lower.getWeight()) && lower.getWeight() != 0) {
+                originalLower.setWeight(lower.getWeight());
+            }
+
+            // color
+            if (Objects.nonNull(lower.getColor()) && !"".equalsIgnoreCase(lower.getColor())) {
+                originalLower.setColor(lower.getColor());
+            }
+
+            // isComplete
+            if (Objects.nonNull(lower.isComplete())) {
+                originalLower.setComplete(lower.isComplete());
+            }
+
+            return lowerRepository.save(originalLower);
+        }
+        return null;
+    }
 
     // DELETE
-    public void deleteLower(int id) {
-        lowerRepository.deleteById(id);
+    public String deleteLower(int id) {
+        if (lowerRepository.findById(id).isPresent()) {
+            lowerRepository.deleteById(id);
+            return "Lower successfully deleted.";
+        }
+        return "Lower not found in database.";
     }
+
 }
